% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLMTranslate-package.R
\docType{package}
\name{LLMTranslate}
\alias{LLMTranslate}
\alias{LLMTranslate-package}
\title{LLMTranslate: Shiny App for TRAPD/ISPOR Survey Translation with LLMs}
\description{
A Shiny application that automates forward translation, optional blind
back-translation, and reconciliation following the TRAPD/ISPOR workflow.
Powered by the ellmer package for provider-agnostic LLM communication.
}
\details{
Use \code{\link[=run_app]{run_app()}} to launch the interactive Shiny app, or
\code{\link[=translate_batch]{translate_batch()}} / \code{\link[=translate_item]{translate_item()}} for programmatic translation.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
LLMTranslate::run_app()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=run_app]{run_app()}}, \code{\link[=translate_batch]{translate_batch()}}, \code{\link[=translate_item]{translate_item()}}
}
\author{
\strong{Maintainer}: Jonas R. Kunst \email{jonas.r.kunst@bi.no} [copyright holder]

Other contributors:
\itemize{
  \item Athanasia M. Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID}) (drmowinckels) [contributor]
}

}
