% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the LLM Survey Translator shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Passed to shiny::runApp.}
}
\value{
No return value, called for side effects (launches a shiny app).
}
\description{
Launch the LLM Survey Translator shiny app
}
\examples{
if (interactive()) {
  run_app()
}
}
